HomeWork № 3 - fix 1
