+++ src/HomeWork/Lesson3/Armoredable.java
@@ -1,0 +1,7 @@
+package HomeWork.Lesson3;
+
+public interface Armoredable {
+  void defend();
+}
+
+
