+++ src/HomeWork/Lesson3/Shootable.java
@@ -1,0 +1,5 @@
+package HomeWork.Lesson3;
+
+public interface Shootable {
+  void shoot();
+}
