+++ src/HomeWork/Lesson3/Sprayable.java
@@ -1,0 +1,5 @@
+package HomeWork.Lesson3;
+
+public interface Sprayable {
+  void sprayLiquid();
+}
