+++ src/HomeWork/Lesson3/Swimable.java
@@ -1,0 +1,5 @@
+package HomeWork.Lesson3;
+
+public interface Swimable {
+  void swim();
+}
